+++ client/sites/overview/components/support-card/index.js
@@ -32,7 +32,7 @@ export default function SupportCard() {
 	const { setShowHelpCenter, setNavigateToRoute } = useDataStoreDispatch( HELP_CENTER_STORE );
 	const { url } = useStillNeedHelpURL();
 
-	const onClick = async () => {
+	const onClick = () => {
 		setNavigateToRoute( url );
 		setShowHelpCenter( true );
 		dispatch( trackNavigateGetHelpClick() );
