+++ packages/help-center/src/components/help-center-chat.tsx
@@ -4,10 +4,7 @@
  */
 import { recordTracksEvent } from '@automattic/calypso-analytics';
 import OdieAssistantProvider, { OdieAssistant } from '@automattic/odie-client';
-import { useCurrentSupportInteraction } from '@automattic/odie-client/src/data/use-current-support-interaction';
-import { useManageSupportInteraction } from '@automattic/odie-client/src/data/use-manage-support-interaction';
-import { isTestModeEnvironment, useCanConnectToZendeskMessaging } from '@automattic/zendesk-client';
-import { useQueryClient } from '@tanstack/react-query';
+import { useCanConnectToZendeskMessaging } from '@automattic/zendesk-client';
 import { useEffect } from '@wordpress/element';
 import { useLocation, useNavigate } from 'react-router-dom';
 import { useHelpCenterContext } from '../contexts/HelpCenterContext';
@@ -35,7 +32,7 @@ export function HelpCenterChat( {
 	const userFieldMessage = params.get( 'userFieldMessage' );
 	const siteUrl = params.get( 'siteUrl' );
 	const siteId = params.get( 'siteId' );
-	const interactionId = params.get( 'id' );
+
 	const { forceEmailSupport } = useChatStatus();
 
 	useEffect( () => {
@@ -47,39 +44,6 @@ export function HelpCenterChat( {
 			navigate( '/' );
 		}
 	}, [ navigate, preventOdieAccess ] );
-
-	const queryClient = useQueryClient();
-	const { startNewInteraction, isMutating: isStartingNewInteraction } =
-		useManageSupportInteraction();
-	const currentInteractionQuery = useCurrentSupportInteraction();
-
-	useEffect( () => {
-		// If a user lands at /odie without an ID, we need to create a new support interaction and redirect to the new URL.
-		if ( ! interactionId && ! isStartingNewInteraction ) {
-			const newID = crypto.randomUUID();
-			startNewInteraction( {
-				event_source: 'help-center',
-				event_external_id: newID,
-			} ).then( ( interaction ) => {
-				navigate( `/odie?id=${ interaction.uuid }`, { replace: true } );
-				const isTestMode = isTestModeEnvironment();
-				const queryKey = [
-					'support-interactions',
-					'get-interaction-by-id',
-					interaction.uuid,
-					isTestMode,
-				];
-				queryClient.setQueryData( queryKey, interaction );
-			} );
-		}
-	}, [
-		interactionId,
-		startNewInteraction,
-		navigate,
-		queryClient,
-		currentInteractionQuery,
-		isStartingNewInteraction,
-	] );
 
 	return (
 		<OdieAssistantProvider
