+++ packages/odie-client/src/index.tsx
@@ -1,8 +1,12 @@
+import { isTestModeEnvironment } from '@automattic/zendesk-client';
+import { useQueryClient } from '@tanstack/react-query';
 import { useEffect } from 'react';
+import { useLocation, useNavigate } from 'react-router-dom';
 import { ClosedConversationFooter } from './components/closed-conversation-footer';
 import { MessagesContainer } from './components/message/messages-container';
 import { OdieSendMessageButton } from './components/send-message-input';
 import { useOdieAssistantContext, OdieAssistantProvider } from './context';
+import { useManageSupportInteraction } from './data';
 import { useCurrentSupportInteraction } from './data/use-current-support-interaction';
 import { hasCSATMessage, interactionHasEnded } from './utils';
 
@@ -11,6 +15,42 @@ import './style.scss';
 export const OdieAssistant: React.FC = () => {
 	const { trackEvent, currentUser, chat } = useOdieAssistantContext();
 	const { data: currentSupportInteraction } = useCurrentSupportInteraction();
+	const { search } = useLocation();
+	const params = new URLSearchParams( search );
+	const interactionId = params.get( 'id' );
+	const queryClient = useQueryClient();
+	const { startNewInteraction, isMutating: isStartingNewInteraction } =
+		useManageSupportInteraction();
+	const currentInteractionQuery = useCurrentSupportInteraction();
+	const navigate = useNavigate();
+
+	useEffect( () => {
+		// If a user lands at /odie without an ID, we need to create a new support interaction and redirect to the new URL.
+		if ( ! interactionId && ! isStartingNewInteraction ) {
+			const newID = crypto.randomUUID();
+			startNewInteraction( {
+				event_source: 'help-center',
+				event_external_id: newID,
+			} ).then( ( interaction ) => {
+				navigate( `/odie?id=${ interaction.uuid }`, { replace: true } );
+				const isTestMode = isTestModeEnvironment();
+				const queryKey = [
+					'support-interactions',
+					'get-interaction-by-id',
+					interaction.uuid,
+					isTestMode,
+				];
+				queryClient.setQueryData( queryKey, interaction );
+			} );
+		}
+	}, [
+		interactionId,
+		startNewInteraction,
+		navigate,
+		queryClient,
+		currentInteractionQuery,
+		isStartingNewInteraction,
+	] );
 
 	const chatHasCSATMessage = hasCSATMessage( chat );
 	const showClosedConversationFooter =
